smaller structures

+++ src/vectors.c
@@ -1,4 +1,4 @@
 #include "vectors.h"
 
-declare_cvector(uint32_t, int) declare_cvector(block_ptr_t, int)
+declare_cvector(uint16_t, int16_t) declare_cvector(block_ptr_t, int16_t)
     declare_cvector(pair2dl_t, long)
